feat: change alert into float

+++ internal/services/alert/alert.go
@@ -19,8 +19,8 @@ type PriceChangesAlert struct {
 	ChangePct       float64 `json:"change_pct"`
 	Name            string  `json:"name"`
 	AlertType       string  `json:"alert_type"`
-	LatestSellPrice int32   `json:"latest_price"`
-	OldSellPrice    int32   `json:"old_price"`
+	LatestSellPrice float64 `json:"latest_price"`
+	OldSellPrice    float64 `json:"old_price"`
 }
 
 func NewAlertService(priceService *price.PriceService, messaging *messaging.Publisher) *AllertService {
